refactor(stella-cli): split the provider listing renderers out of config.rs
config.rs sat at 1498 lines — two under the ratchet — so #1839's ~43 lines
of cache-TTL resolution crossed it. The baseline takes no new entries and a
file approaching the limit gets split, not grown over it, so this moves a
coherent unit out rather than relocating just the new lines.

The seam is concern, not line count: config.rs resolves what THIS session
will use (which provider, which credential, which model); the extracted
block renders what is AVAILABLE. They share only PROVIDERS,
effective_builtin, and custom_provider. Both renderers travel together
because they are two views of one listing — ANSI to stdout for `stella
models`, and a returned String for the deck, whose alternate screen stdout
printing would corrupt — and drift between them would make `/models`
disagree with `stella models`.

No behavior change: the block moved verbatim, only its `use` header is new.
config.rs is back to 1324 lines.
